+++ src/app/sections/payments/index.ts
@@ -1,0 +1,1 @@
+export * from './payments.module';
