--- src/app/sections/search-payments/index.ts
@@ -1,1 +1,0 @@
-export * from './search-payments.module';
